conditionally delete folder in s3 sync

+++ ci/tasks/aws-s3-sync/task.sh
@@ -11,6 +11,9 @@ export CHECK="√"
 
 cd styleguide
 
-aws s3 rm ${destination} --recursive
+if [[ ${destination} == *"s3://"* ]]; then
+  aws s3 rm ${destination} --recursive
+fi
+
 aws s3 sync ${source} ${destination}
 echo -e "${GREEN}${CHECK} Packages deployed from ${source} to ${destination} ${NC}"
